+++ CHANGELOG.rst
@@ -3,11 +3,17 @@
 
 Added
 ^^^^^
+* `@senwu`_: Add micro/macro metric options which have split-wise micro/macro average
+  and global micro/macro average.
 * `@senwu`_: Add contrib folder to support unofficial usages.
 
 Fixed
 ^^^^^
 * `@senwu`_: Add missing pytest.ini and fix test cannot remove temp dir issue.
+
+Changed
+^^^^^^^
+* `@senwu`_: Update the metric to return one metric value by default.
 
 ..
   For convenience, all username links for contributors can be listed here